// peso_lucky_admin-main/src/pages/Report/Bill/data.d.ts
export interface Params {}
export interface ListItem {
  gift_price_detail: ReactNode;
  merchant_gift_price: any;
  commission_detail: ReactNode;
  bill_money: ReactNode;
  pool_amount: ReactNode;
  minimum_charge: ReactNode;
  rate_plan_res: any;
  Commission_money: number;
  id: string;
  merchant_id: number;
  merchant_code: string;
  profit_amount: number;
  flow_amount: number;
  commission_amount: number;
  other_amount: number;
  bill_amount: number;
  tiered_rate: number;
  other_fee: number;
  commission_type: number;
  rate_plan_id: number;
  billTime: string;
  created_at: number;
  updated_at: number;
  created_by: string;
  updated_by: string;
  site_id: number;
  plan_name: string;
}